requirements.txt
# === Runtime ===
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic>=2.7,<3
python-dotenv>=1.0,<2

transformers>=4.44,<5
tokenizers>=0.19,<0.20
safetensors>=0.4,<0.5
sentencepiece>=0.2,<0.3   # XLM-R 계열 토크나이저(필요 시)

# === Testing (선택) ===
pytest>=8
httpx>=0.27

# === PyTorch는 환경에 맞게 별도 설치 ===
# CPU:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# CUDA 12.1:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
